refactor(control-plane): move status collection into status bounded context
